Close virtual keyboard better
Previously, there were three issues:

1) Animating opacity. This is more a matter of preference, but we
believe that as the keyboard slides up from the bottom of the screen, it
should not also transition from opacity 0 to opacity 100.

2) Timing of visibility change. It used to be the case that the keyboard
would immediately disappear instead of sliding out of view.

3) Blur. On our custom virtual keyboard, we have a button for hiding the
keyboard that calls the hideVirtualKeyboard. We need the math field to blur
on this command so that the next time the user taps on the math field, the
keyboard appears (we are showing the keyboard "onfocus").

Test plan:

I modified the virtual keyboard example (but didn't commit that code)
and played around with it to ensure the desired behavior.