Fix issue with spacing around radicals
I don't fully understand what's going on here, but the issue appears to have been introduced in fb3b0d87e6862cefd8e383e59f1235e9469fa035 when changing this check from `if (!atom.index)` to  checking `if (typeof atom.index === 'undefined')`. It seems that `atom.index` is actually `null` in the relevant case, so this change ought to handle the case of `null` or `undefined` but not `0`, which was presumably the intent of the previous change that broke it. I dunno.